recommendation-engine/requirements.txt
pandas==2.2.2
transformers==4.40.1
torch==2.3.0
scikit-learn==1.4.2
flask==3.0.3
python-dotenv
elasticsearch==8.13.0
gunicorn